Fix EtherLab Master package dependencies
The `lcec` component strictly only needs the `libethercat` package to
run; this is added to the `Depends:` list.

The host does need an EtherCAT master, provided by the
`etherlabmaster` package, but in a Docker container the master runs on
the host, and that package doesn't need to be installed inside the
container.  Attempting to install the package, where there is no
kernel package available, fails.